+++ example/android/app/src/main/java/net/jonhanson/flutter_native_splash_example/MainActivity.java
@@ -1,6 +1,0 @@
-package net.jonhanson.flutter_native_splash_example;
-
-import io.flutter.embedding.android.FlutterActivity;
-
-public class MainActivity extends FlutterActivity {
-}