+++ example/android/app/src/main/kotlin/net/jonhanson/flutter_native_splash_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package net.jonhanson.flutter_native_splash_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity()